Create a task class, constructor, and method to add new task to array and ls
